kg_transform/pyproject.toml
[tool.poetry]
name = "kg-transform"
version = "0.1.0"
description = "Knowledge Graph Transformation Tool"
authors = ["Developer"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0"
djangorestframework = "^3.14"
drf-spectacular = "^0.27"
django-environ = "^0.11"
rdflib = "^7.0"
pyld = "^2.0"
django-extensions = "^3.2"
httpx = "^0.27"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
pytest-django = "^4.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "kg_transform.settings"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
